Create abstract formField component
