+++ eisbuk-admin/src/components/atoms/FormSection/index.ts
@@ -1,0 +1,3 @@
+import FormSection from "./FormSection";
+
+export default FormSection;
